+++ Assets/Scripts/GameEngine/Character/Move/Effect/EffectContext.cs
@@ -8,7 +8,7 @@ public class EffectContext
     public Character Source { get; }
     public Character Target { get; }
 
-    private readonly Dictionary<string, int> _results = new();
+    private readonly Dictionary<string, float> _results = new();
 
     public EffectContext(Character source, Character target)
     {
@@ -19,13 +19,13 @@ public class EffectContext
     public Character ResolveTarget(TargetType type)
         => type == TargetType.Self ? Source : Target;
 
-    public void StoreResult(string? id, int value)
+    public void StoreResult(string? id, float value)
     {
         if (id != null)
             _results[id] = value;
     }
 
-    public int GetResult(string id)
+    public float GetResult(string id)
     {
         if (!_results.TryGetValue(id, out var value))
             throw new Exception($"Effect result '{id}' not found");
