+++ Assets/Scripts/GameEngine/Character/Move/Effect/Value/IValue.cs
@@ -1,4 +1,4 @@
 public interface IValue
 {
-    public int GetValue(EffectContext context);
+    public float GetValue(EffectContext context);
 }
